Refactor stash screen
