+++ app/(tabs)/stash.tsx
@@ -7,25 +7,21 @@ import {
   TouchableOpacity,
   Image,
   StyleSheet,
-  Dimensions,
 } from "react-native";
 import { MaterialIcons } from "@expo/vector-icons";
-import { GLView } from "expo-gl";
-import { Renderer } from "expo-three";
-import * as THREE from "three";
-import { GLTFLoader } from "three-stdlib";
 import { WebView } from "react-native-webview";
 import AsyncStorage from "@react-native-async-storage/async-storage";
 import ReadyPlayerMe, { RPMUser } from "@/lib/ReadyPlayerMe";
 import BottomDrawer from "../../components/BottomDrawer";
 import LootCard from "../../components/LootCard";
+import StashHeader from "../../components/StashHeader";
+import GLBModelViewer from "../../components/GLBModelViewer";
 import { lootItems } from "../constants/lootData";
 
 StashScreen.options = {
   headerShown: false,
 };
 
-const { width, height } = Dimensions.get("window");
 
 const DEFAULT_USER = {
   username: "@Ashon",
@@ -173,7 +169,7 @@ export default function StashScreen() {
     <SafeAreaView style={styles.safeContainer}>
       <ScrollView>
         <View style={styles.container}>
-          <Header coins={user.coins} />
+          <StashHeader coins={user.coins} />
           <View style={styles.detailsRow}>
             <View style={styles.avatarRow}>
               <Image
@@ -251,188 +247,10 @@ export default function StashScreen() {
   );
 }
 
-// ... keep Header, GLBModelViewer, and styles the same ...
-
-type HeaderProps = { coins: string };
-
-function Header({ coins }: HeaderProps) {
-  return (
-    <View style={styles.header}>
-      <View style={styles.headerBottom}>
-        <TouchableOpacity style={styles.iconButton}>
-          <MaterialIcons name="notifications" size={24} color="#fff" />
-        </TouchableOpacity>
-        <View style={styles.coinsContainer}>
-          <Image
-            source={require("../../assets/images/credits.png")} // adjust the path
-            style={styles.coinImage}
-          />
-          <Text style={styles.coinText}>{coins}</Text>
-        </View>
-        <TouchableOpacity style={styles.iconButton}>
-          <MaterialIcons name="person" size={24} color="#fff" />
-        </TouchableOpacity>
-      </View>
-    </View>
-  );
-}
-
-type GLBModelViewerProps = { modelUrl: string };
-
-function GLBModelViewer({ modelUrl }: GLBModelViewerProps) {
-  const frame = useRef<number | null>(null);
-  const mixer = useRef<THREE.AnimationMixer | null>(null);
-  const clock = useRef(new THREE.Clock());
-
-  useEffect(() => {
-    return () => {
-      if (frame.current) cancelAnimationFrame(frame.current);
-    };
-  }, []);
-
-  return (
-    <GLView
-      key={modelUrl}
-      style={styles.glView}
-      onContextCreate={async (gl) => {
-        const { drawingBufferWidth: w, drawingBufferHeight: h } = gl;
-
-        const scene = new THREE.Scene();
-
-        // ✅ Subtle vertical gradient background
-        const canvas = document.createElement("canvas");
-        canvas.width = 1;
-        canvas.height = 256;
-        const context = canvas.getContext("2d");
-        const gradient = context.createLinearGradient(0, 0, 0, canvas.height);
-        gradient.addColorStop(0, "#Ffff00"); // lighter gray
-        gradient.addColorStop(1, "#666666"); // darker gray
-        context.fillStyle = gradient;
-        context.fillRect(0, 0, 1, canvas.height);
-        const texture = new THREE.CanvasTexture(canvas);
-        scene.background = texture;
-
-        const camera = new THREE.PerspectiveCamera(50, w / h, 0.1, 1000);
-        camera.position.set(0, 1.6, 4.5);
-        camera.lookAt(0, 1.55, 0);
-
-        const renderer = new Renderer({ gl });
-        renderer.setSize(w, h);
-
-        // ✅ Lighting
-        scene.add(new THREE.AmbientLight(0xffffff, 0.4));
-
-        const keyLight = new THREE.DirectionalLight("#FFFFFF", 0.8);
-        keyLight.position.set(5, 5, 5);
-        scene.add(keyLight);
-
-        const fillLight = new THREE.DirectionalLight("#6794FF", 3);
-        fillLight.position.set(-5, 2, 3);
-        scene.add(fillLight);
-
-        const backLight = new THREE.DirectionalLight("#FFB878", 6);
-        backLight.position.set(0, 2, -5);
-        scene.add(backLight);
-
-        // ✅ Particle system
-        const particleCount = 300;
-        const particlesGeometry = new THREE.BufferGeometry();
-        const positions = new Float32Array(particleCount * 3);
-        for (let i = 0; i < particleCount * 3; i++) {
-          positions[i] = (Math.random() - 0.5) * 10;
-        }
-        particlesGeometry.setAttribute(
-          "position",
-          new THREE.BufferAttribute(positions, 3)
-        );
-        const particlesMaterial = new THREE.PointsMaterial({
-          color: 0xffffff,
-          size: 0.05,
-          transparent: true,
-          opacity: 0.8,
-          depthWrite: false,
-          blending: THREE.AdditiveBlending,
-        });
-        const particles = new THREE.Points(
-          particlesGeometry,
-          particlesMaterial
-        );
-        scene.add(particles);
-
-        // ✅ Load GLB character
-        const loader = new GLTFLoader();
-        loader.load(modelUrl, (gltf) => {
-          const model = gltf.scene;
-          model.scale.set(2, 2, 2);
-          model.position.set(0, -0.2, 0);
-          scene.add(model);
-
-          mixer.current = new THREE.AnimationMixer(model);
-
-          // ✅ Load animation
-          const animLoader = new GLTFLoader();
-          animLoader.load(
-            "https://raw.githubusercontent.com/readyplayerme/animation-library/master/feminine/glb/idle/F_Standing_Idle_Variations_001.glb",
-            (animGltf) => {
-              if (animGltf.animations.length > 0 && mixer.current) {
-                const action = mixer.current.clipAction(animGltf.animations[0]);
-                action.play();
-              }
-            }
-          );
-        });
-
-        // ✅ Animation loop
-        const animate = () => {
-          frame.current = requestAnimationFrame(animate);
-          const delta = clock.current.getDelta();
-
-          if (mixer.current) {
-            mixer.current.update(delta);
-          }
-
-          particles.rotation.y += delta * 0.1;
-
-          renderer.render(scene, camera);
-          gl.endFrameEXP();
-        };
-        animate();
-
-        // ✅ Cleanup on unmount
-        return () => {
-          cancelAnimationFrame(frame.current);
-          renderer.dispose();
-          particlesGeometry.dispose();
-          particlesMaterial.dispose();
-          texture.dispose();
-          gl = null;
-        };
-      }}
-    />
-  );
-}
 
 const styles = StyleSheet.create({
   safeContainer: { flex: 1, backgroundColor: "#fff" },
   container: { flex: 1, backgroundColor: "#f3f4f6" },
-  header: { backgroundColor: "#3b82f6", padding: 16 },
-  headerBottom: {
-    flexDirection: "row",
-    justifyContent: "space-between",
-    alignItems: "center",
-    marginTop: 12,
-  },
-  iconButton: { padding: 4 },
-  coinsContainer: {
-    flexDirection: "row",
-    alignItems: "center",
-    backgroundColor: "#2563eb",
-    borderRadius: 9999,
-    paddingHorizontal: 8,
-    paddingVertical: 4,
-  },
-  coinImage: { width: 32, height: 32, marginRight: 4 },
-  coinText: { color: "#fff", fontSize: 20, fontWeight: "bold" },
   detailsRow: {
     flexDirection: "row",
     justifyContent: "space-between",
@@ -457,7 +275,6 @@ const styles = StyleSheet.create({
   },
   detailsButtonText: { color: "#fff", fontSize: 12, fontWeight: "bold" },
   modelWrapper: { alignItems: "center", marginTop: 16 },
-  glView: { width: width, height: height * 0.5 },
   dots: { flexDirection: "row", justifyContent: "center", marginTop: 8 },
   dot: {
     width: 8,
